agent/requirements.txt
# Core AI Framework
pydantic-ai[openai]>=0.0.14
pydantic>=2.9.2

# LLM Integration
httpx>=0.27.2
openai>=1.54.4

# Memory Management
# Using stable PyPI release (1.0.0) with pgvector for PostgreSQL vector storage
mem0ai==1.0.0
psycopg[pool]>=3.1.0
ollama>=0.1.0
neo4j>=5.14.0
langchain-neo4j>=0.1.0
rank-bm25>=0.2.2

# Knowledge Graph Memory (Graphiti)
graphiti-core>=0.5.0

# Observability
langfuse>=3.8.1

# Web Framework (for API)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.1.0
rich>=13.9.4
ntplib>=0.4.0
pytz>=2024.1
